applied authentication to all the routes that needed to be protected

+++ api/app.js
@@ -29,6 +29,23 @@ app.use(function(req, res, next) {
     next();
 });
 
+// check whether the request has a valid JWT access token 
+let authenticate = (req, res, next) => {
+    let token = req.header('x-access-token');
+
+    // verify the JWT 
+    jwt.verify(token, User.getJWTSecret(), (err, decoded) => {
+        if (err) {
+            // jwt is invalid - * DO NOT AUTHENTICATE*
+            res.status(401).send(err);
+        } else {
+            // jwt is valid 
+            req.user_id = decoded._id;
+            next();
+        }
+    });
+}
+
 // verify refresh token middleware (which will be erifying the session)
 let verifySession = (req, res, next) => {
     // grab the refresh token from the request header 
@@ -89,10 +106,14 @@ let verifySession = (req, res, next) => {
  * GET /lists
  * Purpose: Get all lists
  */
-app.get('/lists', (req, res) => {
-    // return an array of all lists in the database
-    List.find().then((lists) => {
+app.get('/lists', authenticate, (req, res) => {
+    // return an array of all lists that belongs to the authenticated user
+    List.find({
+        _userId: req.user_id
+    }).then((lists) => {
         res.send(lists);
+    }).catch((e) => {
+        res.send(e);
     });
 });
 
@@ -100,13 +121,14 @@ app.get('/lists', (req, res) => {
  * POSTS /lists
  * Purpose: Create a list
  */
-app.post('/lists', (req, res) => {
+app.post('/lists', authenticate, (req, res) => {
     // create a new list and return the new list document back to the user (which includes the id)
     // The list information (fields) will be passed in via JSON request body
     let title = req.body.title;
 
     let newList = new List({
-        title
+        title,
+        _userId: req.user_id
     });
     newList.save().then((listDoc) => {
         res.send(listDoc);
@@ -117,11 +139,12 @@ app.post('/lists', (req, res) => {
  * PATCH /lists/:id
  * Purpose: Update a specified list
  */
-app.patch('/lists/:id', (req, res) => {
+app.patch('/lists/:id', authenticate, (req, res) => {
     // update specified list (list document with id in the URL) with new values specified in the JSON body request
-    List.findOneAndUpdate({ _id: req.params.id }, {
-        $set: req.body
-    }).then(() => {
+    List.findOneAndUpdate({ 
+        _id: req.params.id,
+        _userId: req.user_id
+    }, { $set: req.body }).then(() => {
         res.sendStatus(200);
     });
 });
@@ -130,12 +153,17 @@ app.patch('/lists/:id', (req, res) => {
  * DELETE /lists/:id
  * Purpose: Delete a list
  */
-app.delete('/lists/:id', (req, res) => {
+app.delete('/lists/:id', authenticate, (req, res) => {
     // delete a specified list (document with id in the URL)
     List.findOneAndRemove({ 
-        _id: req.params.id
+        _id: req.params.id,
+        _userId: req.user_id
     }).then((removedListDoc) => {
         res.send(removedListDoc);
+
+        // delete all the tasks are in deleted list 
+        deleteTasksFromList(removedListDoc._id);
+
     });
 });
 
@@ -149,36 +177,56 @@ app.delete('/lists/:id', (req, res) => {
  * GET /lists/:listId/tasks
  * Purpose: Get all tasks in a specific list
  */
-app.get('/lists/:listId/tasks', (req, res) => {
+app.get('/lists/:listId/tasks', authenticate, (req, res) => {
     // return all tasks that belong to a specific list specified by listId
     Task.find({
-        _listId: req.params.listId
+        _listId: req.params.listId,
+        _userId: req.user_id
     }).then((tasks) => {
         res.send(tasks);
     });
 });
 
-app.get('/lists/:listId/tasks/:taskId', (req, res) => {
-    Task.findOne({
-        _id: req.params.taskId,
-        _listId: req.params.listId
-    }).then((task) => {
-        res.send(task);
-    });
-});
+// app.get('/lists/:listId/tasks/:taskId', authenticate, (req, res) => {
+//     Task.findOne({
+//         _id: req.params.taskId,
+//         _listId: req.params.listId,
+//         _userId: req.user_id
+//     }).then((task) => {
+//         res.send(task);
+//     });
+// });
 
 /**
  * POST /lists/:listId/tasks
  * Purpose: Create a task in a specific list
  */
-app.post('/lists/:listId/tasks', (req, res) => {
+app.post('/lists/:listId/tasks', authenticate, (req, res) => {
     // create anew task in a list specified by listId
-    let newTask = new Task({
-        title: req.body.title,
-        _listId: req.params.listId
-    });
-    newTask.save().then((taskDoc) => {
-        res.send(taskDoc);
+
+    List.findOne({
+        _id: req.params.listId,
+        _userId: req.user_id
+    }).then((list) => {
+        if (list) {
+            // list object with the specified conditions was found
+            // therefore the currently authenticated user can create new tasks 
+            return true;
+        }
+        // list object is undefined 
+        return false;
+    }).then((canCreateTask) => {
+        if (canCreateTask) {
+            let newTask = new Task({
+                title: req.body.title,
+                _listId: req.params.listId
+            });
+            newTask.save().then((taskDoc) => {
+                res.send(taskDoc);
+            });
+        } else {
+            res.sendStatus(404);
+        }
     });
 });
 
@@ -186,15 +234,34 @@ app.post('/lists/:listId/tasks', (req, res) => {
  * PATCH /lists/:listId/tasks/:taskId
  * Purpose: Create a task in a specific list
  */
-app.patch('/lists/:listId/tasks/:taskId', (req, res) => {
+app.patch('/lists/:listId/tasks/:taskId', authenticate, (req, res) => {
     // update an existing task specified by taskId
-    Task.findOneAndUpdate({ 
-        _id: req.params.taskId,
-        _listId: req.params.listId
-    }, {
-        $set: req.body
-    }).then(() => {
-        res.send({message: 'Updated Successfully'});
+
+    List.findOne({
+        _id: req.params.listId,
+        _userId: req.user_id
+    }).then((list) => {
+        if (list) {
+            // list object with the specified conditions was found
+            // therefore the currently authenticated user can update tasks within this list 
+            return true;
+        }
+        // list object is undefined 
+        return false;
+    }).then((canUpdateTasks) => {
+        if (canUpdateTasks) {
+            // the currently authenticated user can update tasks
+            Task.findOneAndUpdate({ 
+                _id: req.params.taskId,
+                _listId: req.params.listId
+            }, {
+                $set: req.body
+            }).then(() => {
+                res.send({message: 'Updated Successfully'});
+            });
+        } else {
+            res.sendStatus(404);
+        }
     });
 });
 
@@ -202,13 +269,31 @@ app.patch('/lists/:listId/tasks/:taskId', (req, res) => {
  * DELETE /lists/:listId/tasks/:taskId
  * Purpose: Delete a task
  */
-app.delete('/lists/:listId/tasks/:taskId', (req, res) => {
+app.delete('/lists/:listId/tasks/:taskId', authenticate, (req, res) => {
     // delete a task specified by taskId
-    Task.findOneAndRemove({
-        _id: req.params.taskId,
-        _listId: req.params.listId
-    }).then((removedTaskDoc) => {
-        res.send(removedTaskDoc);
+    List.findOne({
+        _id: req.params.listId,
+        _userId: req.user_id
+    }).then((list) => {
+        if (list) {
+            // list object with the specified conditions was found
+            // therefore the currently authenticated user can delete tasks within this list
+            return true;
+        }
+        // else - the list object is undefined
+        return false;
+    }).then((canDeleteTasks) => {
+        
+        if (canDeleteTasks) {
+            Task.findOneAndRemove({
+                _id: req.params.taskId,
+                _listId: req.params.listId
+            }).then((removedTaskDoc) => {
+                res.send(removedTaskDoc);
+            })
+        } else {
+            res.sendStatus(404);
+        }
     });
 });
 
@@ -288,6 +373,16 @@ app.get('/users/me/access-token', verifySession, (req, res) => {
         res.status(400).send(e);
     });
 });
+
+
+// *** helper methods ***
+let deleteTasksFromList = (_listId) => {
+    Task.deleteMany({
+        _listId
+    }).then(() => {
+        console.log('Tasks from ' + _listId + ' were deleted');
+    });
+}
 
 app.listen(3000, () => {
     console.log('Server is running on port 3000');
